Added client cert list for SSL Authentication
As part of rsyslog encryption (handled by rsyslog-gnutls) a client
certificate is required for SSL communication.
This adds the "Current User" personal store as a client cert collection
for the SSL authentication. Any acceptible certificate from that store
will then be presented to the server (if required).

Also changed the TcpStream Poll to use SelectWrite, as this poll now
correctly checks that the stream is open and can be written to.